LPS-171473 Inlining class
