--- modules/apps/headless/headless-builder/headless-builder-api/src/main/resources/com/liferay/headless/builder/dto/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
